Replace the data type of `ctx_category_id`, `ctx_scheme_id`, `biz_ctx_id`, `module_id`, etc. to char(36) for UUID.

+++ score-service/src/main/java/org/oagi/score/service/businesscontext/ContextSchemeService.java
@@ -77,7 +77,7 @@ public class ContextSchemeService {
     public boolean hasSameCtxScheme(ContextScheme contextScheme) {
         Condition idMatch = trueCondition();
         if (contextScheme.getContextSchemeId() != null) {
-            idMatch = CTX_SCHEME.CTX_SCHEME_ID.notEqual(ULong.valueOf(contextScheme.getContextSchemeId()));
+            idMatch = CTX_SCHEME.CTX_SCHEME_ID.notEqual(contextScheme.getContextSchemeId());
         }
 
         return dslContext.selectCount().from(CTX_SCHEME).where(
@@ -91,7 +91,7 @@ public class ContextSchemeService {
     public boolean hasSameCtxSchemeName(ContextScheme contextScheme) {
         Condition idMatch = trueCondition();
         if (contextScheme.getContextSchemeId() != null) {
-            idMatch = CTX_SCHEME.CTX_SCHEME_ID.notEqual(ULong.valueOf(contextScheme.getContextSchemeId()));
+            idMatch = CTX_SCHEME.CTX_SCHEME_ID.notEqual(contextScheme.getContextSchemeId());
         }
         return dslContext.selectCount().from(CTX_SCHEME).where(
                 and(CTX_SCHEME.SCHEME_ID.eq(contextScheme.getSchemeId()),
